Added blacklisting
Blacklisting is now available through the launcher, and can r. Whenever samples are successfully analysed, their sample names will be added to the config/blasklist file, which in turn will ensure that these samples are ignored upon the next execution. Currently, a snakemake error shows up, if there are no new samples, while blacklisting is enabled. However, this is because python envokes the exit command (which is also described in the messages). Can potentially be fixed by adding a checkpoint somewhere.